+++ app/src/main/java/rocks/itsnotrocketscience/bejay/api/retrofit/Events.java
@@ -26,7 +26,7 @@ public interface Events {
     Observable<Event> get(@Path("id") int id);
 
     @GET("/events/public_nearby_events/")
-    Observable<ArrayList<Event>> publicNearbyEvents(@Query("lat") long lat, @Query("lng") long lng);
+    Observable<ArrayList<Event>> publicNearbyEvents(@Query("lat") double lat, @Query("lng") double lng);
 
     @GET("/events/search_events/")
     Observable<ArrayList<Event>> searchEvents(@Query("search_term") String term);
